Fixed a thing or two
