+++ src/main/java/Interface_and_Adapters/StartUpScreens/LoginScreen.java
@@ -46,7 +46,7 @@ public class LoginScreen extends JPanel {
             @Override
             public void actionPerformed(ActionEvent e) {
                 Main main = new Main();
-                main.switchPanel(mainPanel, "FIRST");
+                //main.switchPanel(mainPanel, "FIRST"); generates error
             }
 
         });
@@ -64,7 +64,7 @@ public class LoginScreen extends JPanel {
                             username.getText(), String.valueOf(password.getPassword()));
                     Main main = new Main();
                     mainPanel.add(new TabPanel(mainPanel, account), "FOURTH");
-                    main.switchPanel(mainPanel, "FOURTH");
+                    //main.switchPanel(mainPanel, "FOURTH"); generates error
 
 
                 } catch (Exception ex) {
